+++ src/Core/Application/UseCase/GetGenerations/GetGenerationUseCaseInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Core\Application\UseCase\GetGenerations;
+
+interface GetGenerationUseCaseInterface
+{
+    public function execute(GetGenerationsPresenterInferface $presenter): void;
+}